% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{clinical_settings}
\alias{clinical_settings}
\title{Settings}
\usage{
clinical_settings(timeline, ...)
}
\arguments{
\item{timeline}{An object of class \code{clinical_timeline}
as returned by \code{\link[=clinical_timeline]{clinical_timeline()}}.}

\item{...}{Settings, see
\href{https://github.com/RhoInc/clinical-timelines/wiki/Configuration}{official documentation}}
}
\description{
Pass settings.
}
