% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_timeline.R
\name{clinical_timeline}
\alias{clinical_timeline}
\title{Clinical Timeline}
\usage{
clinical_timeline(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{Dataset to visualise.}

\item{width, height}{Dimensions.}

\item{elementId}{HTML ID to use.}
}
\description{
Create a clinical timeline.
}
